+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # []
 
 - Fixed suboptimal layout in some cases
+- Fixed entity GLM condition list not complete
 
 # [1.6.0]